+++ example_async/lib/config/avoid_mixed_environments_fixture.dart
@@ -1,0 +1,6 @@
+// ignore_for_file: unused_local_variable, unused_element
+
+/// Fixture for `avoid_mixed_environments` lint rule.
+
+// TODO: Add bad/good examples for avoid_mixed_environments
+void main() {}