+++ example_async/lib/debug/avoid_print_in_release_fixture.dart
@@ -1,0 +1,6 @@
+// ignore_for_file: unused_local_variable, unused_element
+
+/// Fixture for `avoid_print_in_release` lint rule.
+
+// TODO: Add bad/good examples for avoid_print_in_release
+void main() {}