+++ example_packages/lib/freezed/prefer_freezed_default_values_fixture.dart
@@ -1,0 +1,6 @@
+// ignore_for_file: unused_local_variable, unused_element
+
+/// Fixture for `prefer_freezed_default_values` lint rule.
+
+// TODO: Add bad/good examples for prefer_freezed_default_values
+void main() {}