+++ example_packages/lib/freezed/require_freezed_explicit_json_fixture.dart
@@ -1,0 +1,6 @@
+// ignore_for_file: unused_local_variable, unused_element
+
+/// Fixture for `require_freezed_explicit_json` lint rule.
+
+// TODO: Add bad/good examples for require_freezed_explicit_json
+void main() {}