+++ example_packages/lib/graphql/avoid_graphql_string_queries_fixture.dart
@@ -1,0 +1,6 @@
+// ignore_for_file: unused_local_variable, unused_element
+
+/// Fixture for `avoid_graphql_string_queries` lint rule.
+
+// TODO: Add bad/good examples for avoid_graphql_string_queries
+void main() {}