+++ example_platforms/lib/platforms/avoid_x11_only_assumptions_fixture.dart
@@ -1,0 +1,6 @@
+// ignore_for_file: unused_local_variable, unused_element
+
+/// Fixture for `avoid_x11_only_assumptions` lint rule.
+
+// TODO: Add bad/good examples for avoid_x11_only_assumptions
+void main() {}